NEW Create DBClassNameVarchar
